==> code/1.bankruptcy_data_collection.ipynb <==
import pandas as pd
import numpy as np
import os
import requests
import json
from tqdm import tqdm

# 1. Collect the latest 10-k file dates for all the companies

cik = []
sic = []
sic_description = []
latest_filedate = []
directory = '../data/submissions'
for filename in tqdm(os.listdir(directory)):
    if re.match(r'^CIK\d{10}.json$', filename):
        file_path = os.path.join(directory, filename)
        with open(file_path, 'r') as file:
            submission = json.load(file)
            for i in range(len(submission['filings']['recent']['form'])):
                if submission['filings']['recent']['form'][i] == '10-K':
                    cik.append(submission['cik'])
                    sic.append(submission['sic'])
                    sic_description.append(submission['sicDescription'])
                    latest_filedate.append(submission['filings']['recent']['filingDate'][i])
                    break 

latest10k = pd.DataFrame({
    'cik': cik,
    'sic': sic,
    'sic_description': sic_description,
    'latest_filedate': latest_filedate
})
latest10k

## 1.1. Filtering onbankruptcy Data

latest10k['latest_filedate'] = pd.to_datetime(latest10k['latest_filedate'])
latest10k["cik"] = latest10k["cik"].astype(str).str.zfill(10)

# select companies with 10-k files in 2023 or later
nonbank_list = latest10k[latest10k['latest_filedate'].dt.year >= 2023]["cik"]

# Extract all the 10-k files for those companies
directory = "../data/submissions"
cik_list = []
sic_list = []
sic_description_list = []
filedate_list = []
accession_num_list = []
primary_doc_list = []

for cik in tqdm(nonbank_list):
    filename = "CIK" + cik + ".json"
    file_path = os.path.join(directory, filename)
    with open(file_path, 'r') as file:
        submission = json.load(file)
        for i in range(len(submission['filings']['recent']['form'])):
            if submission['filings']['recent']['form'][i] == '10-K':
                cik_list.append(submission['cik'])
                sic_list.append(submission['sic'])
                sic_description_list.append(submission['sicDescription'])
                filedate_list.append(submission['filings']['recent']['filingDate'][i])
                accession_num_list.append(submission['filings']['recent']['accessionNumber'][i])
                primary_doc_list.append(submission['filings']['recent']['primaryDocument'][i])

nonbank = pd.DataFrame({
    'cik': cik_list,
    'sic': sic_list,
    'sic_description': sic_description_list,
    'filedate': filedate_list,
    'accession_num': accession_num_list,
    'primary_doc': primary_doc_list
}, dtype = str)
nonbank['filelink'] = 'https://www.sec.gov/Archives/edgar/data/' +\
    nonbank['cik'] + '/' + nonbank['accession_num'].str.replace('-', '') + '/' + nonbank['primary_doc']

nonbank.to_csv('../data/nonbankrupt.csv')

## Bankruptcy data

# Load BRD data
brd_data = pd.read_csv("../data/florida_ucla_brd.csv", encoding='unicode_escape')
# select items with non-empty cik and 10k file links.
filtered_data = brd_data.dropna(subset=['CikBefore', 'Date10kBefore'])
bankrupt_data = filtered_data[['CikBefore', 'Date10kBefore']]
bankrupt_data['CikBefore'] = bankrupt_data['CikBefore'].astype('int')
brd_data['Date10kBefore'] = pd.to_datetime(brd_data['Date10kBefore'])
brd_data_2009 = brd_data[brd_data['Date10kBefore'].dt.year >= 2009]

latest10k_2009 = latest10k[latest10k["latest_filedate"].dt.year >= 2009]
latest10k_2009['cik'] = latest10k_2009['cik'].astype('int64')

result = pd.merge(brd_data_2009, latest10k_2009, left_on = 'CikBefore', right_on = 'cik', how = 'inner')

accession_num_list = []
primary_doc_list = []
directory = '../data/submissions'
for i in range(result.shape[0]):
    filename = "CIK" + result['cik'].astype(str).str.zfill(10)[i] + ".json"
    file_path = os.path.join(directory, filename)
    with open(file_path, 'r') as file:
        submission = json.load(file)
        for j in range(len(submission['filings']['recent']['form'])):
            if submission['filings']['recent']['form'][j] == '10-K' and submission['filings']['recent']['filingDate'][j] == result['latest_filedate'].astype(str)[i]:
                accession_num_list.append(submission['filings']['recent']['accessionNumber'][j])
                primary_doc_list.append(submission['filings']['recent']['primaryDocument'][j])
                break

bankruptcy_data = pd.DataFrame({
    'cik': result['cik'],
    'sic': result['sic'],
    'sic_description': result['sic_description'],
    'filedate': result['latest_filedate'],
    'accession_num': accession_num_list,
    'primary_doc': primary_doc_list
}, dtype = str)

bankruptcy_data['filelink'] = 'https://www.sec.gov/Archives/edgar/data/' + bankruptcy_data['cik'] + '/' + bankruptcy_data['accession_num'].str.replace('-', '') + '/' + bankruptcy_data['primary_doc']
bankruptcy_data = bankruptcy_data.reset_index(drop = True)
bankruptcy_data

bankruptcy_data.to_csv('../data/bankruptc.csv')
