==> code/2.1.merge_data.ipynb <==
import zipfile
import pandas as pd
import numpy as np
import os
import json
from tqdm import tqdm

# Load bankruptcy and non-bankruptcy data 

bank = pd.read_csv("../data/bankrupt.csv", index_col = 0)
bank["bank_status"] = 1
nonbank = pd.read_csv("../data/nonbankrupt.csv", index_col = 0)
nonbank["bank_status"] = 0
all_company = pd.concat([bank, nonbank], ignore_index = True)
all_company["cik"] = all_company["cik"].astype(str).str.zfill(10)
all_company['sic'] = all_company['sic'].astype(str)

print(nonbank.shape)
print(bank.shape)
print(all_company.shape)

# Collect financial terms

# find financial number from the companfacts folder 
# Path to your large ZIP file
zip_file_path = '../data/companyfacts.zip'
# Open the ZIP file in read mode
with zipfile.ZipFile(zip_file_path, 'r') as zip_ref:
    zip_contents = set(zip_ref.namelist())
    for value in tqdm(all_company["cik"]):
        # Extract a specific file
        specific_file = 'CIK' + value + '.json'
        if specific_file in zip_contents:
            zip_ref.extract(specific_file, '../data/all_company_data')

def collect(company_json_file, finantial_terms, cik, company_facts):
    tmp = pd.DataFrame(columns = ["filed_date"] + finantial_terms)
    for term in finantial_terms:
        if term in company_json_file['facts']['us-gaap']:
            if term == 'CommonStockSharesIssued':
                units = company_json_file['facts']['us-gaap'][term].get("units", {}).get("shares", [])
            else:
                units = company_json_file['facts']['us-gaap'][term].get("units", {}).get("USD", [])
            for unit in units:
                if unit["form"] == "10-K":
                    filed_date = unit.get("filed")
                    if filed_date in tmp["filed_date"].values:
                        tmp.loc[tmp["filed_date"] == filed_date, term] = unit.get("val")
                    else:
                        tmp = pd.concat([tmp, pd.DataFrame({"filed_date": [filed_date], term: [unit.get("val")]})])
    tmp["cik"] = cik
    company_facts = pd.concat([company_facts, tmp])
    return company_facts

def collect_company_facts(finantial_terms):
    folder_path = '../data/all_company_data'
    company_facts = pd.DataFrame(columns = ["cik", "filed_date"] + finantial_terms)
    
    for file_name in tqdm(os.listdir(folder_path)):
        json_file_path = os.path.join(folder_path, file_name)
        with open(json_file_path, 'r') as json_file:
            company_json_file = json.load(json_file)
            if 'us-gaap' not in company_json_file['facts']:
                continue
            cik = company_json_file.get("cik", "Unknown")
            company_facts = collect(company_json_file, finantial_terms, cik, company_facts)
    
    return company_facts

finantial_terms = ['Assets', 'AssetsCurrent', 'Liabilities', 'LiabilitiesCurrent', \
                   'StockholdersEquity', 'CashAndCashEquivalentsAtCarryingValue', \
                   'NetCashProvidedByUsedInOperatingActivities', \
                   'NetCashProvidedByUsedInFinancingActivities', \
                   'NetCashProvidedByUsedInInvestingActivities', \
                   'NetIncomeLoss', 'OperatingIncomeLoss', \
                   'CommonStockValue', 'CommonStockSharesIssued'
                   ]
company_facts = collect_company_facts(finantial_terms)
company_facts["cik"] = company_facts["cik"].astype(str).str.zfill(10)

# Merge financial terms and companies

merged_data = pd.merge(all_company, company_facts, left_on = ['cik', 'filedate'], right_on = ['cik', 'filed_date'], how = 'inner')
merged_data.drop(columns = ['filed_date'], inplace = True)

np.sum(merged_data.isna())

# Drop lines with NA
merged_data = merged_data.dropna()
print(merged_data.shape)
print("The number of bankruptcy companies:", np.sum(merged_data['bank_status'] == 1))

# Drop lines with any finantial term = 0
merged_data = merged_data[(merged_data[finantial_terms] != 0).all(axis = 1)]
print(merged_data.shape)

# Calculate finantial ratios
# liquidity ratios
merged_data['current_ratio'] = merged_data['AssetsCurrent'] / merged_data['LiabilitiesCurrent']

# solvency ratios （有看到用debt的，也有看到用liabilities，因为没有debt只能用liabilities）
merged_data['debt_to_equity'] = merged_data['Liabilities'] / merged_data['StockholdersEquity']
merged_data['debt_to_total_assets'] = merged_data['Liabilities'] / merged_data['Assets']

# profitability ratio (分母要用两年的平均值，算的时候只用了当年的)
merged_data['return_on_assets'] = merged_data['NetIncomeLoss'] / merged_data['Assets']
merged_data['return_on_equity'] = merged_data['NetIncomeLoss'] / merged_data['StockholdersEquity']

# other ratios commonly used in previous research （这些似乎不是常见的ratio，pre的时候可能不好讲这些ratios背后的含义，如果跑模型的时候发现这些用处不大可以删掉，如果有用的话pre的时候要么就说是我们查阅文献得到的）
merged_data['ebit_over_assets'] = merged_data['OperatingIncomeLoss'] / merged_data['Assets']
merged_data['current_assets_over_assets'] = merged_data['AssetsCurrent'] / merged_data['Assets']
merged_data['cash_over_assets'] = merged_data['CashAndCashEquivalentsAtCarryingValue'] / merged_data['Assets']
merged_data['cash_operation_over_assets'] = merged_data['NetCashProvidedByUsedInOperatingActivities'] / merged_data['Assets']
merged_data['cash_operation_over_liabilities'] = merged_data['NetCashProvidedByUsedInOperatingActivities'] / merged_data['Liabilities']


print("The final number of bankruptcy companies:", np.sum(merged_data['bank_status'] == 1))

# Get main industry by grouping the sic code
def map_sic_to_industry(sic_code):
    if 100 <= sic_code <= 999:
        return 'Agriculture, Forestry and Fishing'
    elif 1000 <= sic_code <= 1499:
        return 'Mining'
    elif 1500 <= sic_code <= 1799:
        return 'Construction'
    # Note: 1800-1999 is not used
    elif 2000 <= sic_code <= 3999:
        return 'Manufacturing'
    elif 4000 <= sic_code <= 4999:
        return 'Transportation, Communications, Electric, Gas and Sanitary service'
    elif 5000 <= sic_code <= 5199:
        return 'Wholesale Trade'
    elif 5200 <= sic_code <= 5999:
        return 'Retail Trade'
    elif 6000 <= sic_code <= 6799:
        return 'Finance, Insurance and Real Estate'
    elif 7000 <= sic_code <= 8999:
        return 'Services'
    elif 9100 <= sic_code <= 9729:
        return 'Public Administration'
    elif 9900 <= sic_code <= 9999:
        return 'Nonclassifiable'
    else:
        return 'Unknown'  # For SIC codes that don't fall into any category or missing


merged_data['main_industry'] = merged_data['sic'].astype(float).apply(map_sic_to_industry)

merged_data['filedate'] = pd.to_datetime(merged_data['filedate']) 
merged_data['year'] = merged_data['filedate'].dt.year
merged_data

merged_data.to_csv("../data/data.csv", index = False)
