==> code/2.2.text_cleaning.ipynb <==
import pandas as pd

import re 

from tqdm import tqdm, tqdm_pandas
tqdm.pandas()

import nltk
from nltk.corpus import stopwords
stop_words = set(stopwords.words('english'))
from nltk.stem.snowball import SnowballStemmer
st = SnowballStemmer('english')

data = pd.read_csv("../../project/data/train_test_set.csv")

x = data.loc[0, "clean_items"]
re.sub('[^a-zA-Z]', ' ', x)

for col in ['item_1A', 'item_2', 'item_7']:

    data[col] = data[col].progress_apply(lambda x: x.split('\n', 1)[-1])
    
    # change to lower and remove spaces on either side
    data[col] = data[col].progress_apply(lambda x: x.lower().strip())

    # remove extra spaces in between
    data[col] = data[col].progress_apply(lambda x: re.sub(' +', ' ', x))

    # remove punctuation
    data[col] = data[col].progress_apply(lambda x: re.sub('[^a-zA-Z]', ' ', x))

    # remove stopwords and get the stem
    data[col] = data[col].progress_apply(lambda x: ' '.join(st.stem(text) for text in x.split() if text not in stop_words))

data["clean_items"] = data[['item_1A', 'item_2', 'item_7']].apply(lambda row: ' '.join(row.dropna().astype(str)), axis=1)

data.to_csv('../../project/data/train_test_set.csv', index=False)
