==> code/5.lsa_xgbod.ipynb <==
import pandas as pd

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.svm import SVC
from sklearn.metrics import average_precision_score, classification_report
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import GridSearchCV
from sklearn.decomposition import TruncatedSVD
from sklearn.decomposition import NMF, LatentDirichletAllocation, MiniBatchNMF


from scipy import sparse

import numpy as np
from tqdm import tqdm
from time import time

from imblearn.ensemble import EasyEnsembleClassifier
from imblearn.over_sampling import SMOTE
from imblearn.under_sampling import RandomUnderSampler
from imblearn.pipeline import Pipeline

from pyod.utils.data import evaluate_print
from pyod.models.xgbod import XGBOD 
from pyod.models.ecod import ECOD
from pyod.models.dif import DIF

import lightgbm as lgb

from joblib import dump, load

from nltk.corpus import stopwords
stop_words = set(stopwords.words('english'))
from nltk.stem.snowball import SnowballStemmer
st = SnowballStemmer('english')

# Data Preparation

data = pd.read_csv('../data/train_test_set.csv')
data = pd.get_dummies(data, columns = ['main_industry', 'year'])

X_train = data[data['train'] == 1].drop(columns=['cik', 'sic', 'sic_description', 'filedate', 'accession_num',
       'primary_doc', 'filelink', 'bank_status', 'train'])
X_test = data[data['train'] == 0].drop(columns=['cik', 'sic', 'sic_description', 'filedate', 'accession_num',
       'primary_doc', 'filelink', 'bank_status', 'train'])
y_train = data[data['train']==1]['bank_status']
y_test = data[data['train']==0]['bank_status']

# Standardization for numerical features 
scaler = StandardScaler().fit(X_train.iloc[:, 0:23])
scaled_terms_train = pd.DataFrame(scaler.transform(X_train.iloc[:, 0:23]), columns = X_train.columns[0:23])
X_num_train = pd.concat([scaled_terms_train, X_train.iloc[:,27:].reset_index(drop = True)], axis = 1)
scaled_terms_test = pd.DataFrame(scaler.transform(X_test.iloc[:, 0:23]), columns = X_test.columns[0:23])
X_num_test = pd.concat([scaled_terms_test, X_test.iloc[:,27:].reset_index(drop = True)], axis = 1)

X_text_train = X_train['clean_items']
X_text_test = X_test['clean_items']

# tf-idf features for text features 
vectorizer = TfidfVectorizer(
    sublinear_tf=True, max_df=0.5, min_df=5, stop_words="english"
)

X_text_train = vectorizer.fit_transform(X_text_train)
X_text_test = vectorizer.transform(X_text_test)

svd = TruncatedSVD(n_components=16, random_state=42)
X_pca_train = svd.fit_transform(X_text_train)
X_pca_test = svd.transform(X_text_test)

with open("../res/lsa16.npy", "wb") as f:
    np.save(f, X_pca_train)
    np.save(f, X_pca_test)

# load the pretrained model 
# clf_text = load('../res/xgbod.joblib')

clf_name = 'XGBOD'
clf_text = XGBOD(random_state=42)
clf_text.fit(X_pca_train, y_train)

# get the prediction labels and outlier scores of the training data
y_train_pred = clf_text.labels_  # binary labels (0: inliers, 1: outliers)
y_train_scores = clf_text.decision_scores_  # raw outlier scores

# get the prediction on the test data
y_test_pred = clf_text.predict(X_pca_test)  # outlier labels (0 or 1)
y_test_scores = clf_text.decision_function(X_pca_test)  # outlier scores

# evaluate and print the results
print("\nOn Training Data:")
evaluate_print(clf_name, y_train, y_train_scores)
print("\nOn Test Data:")
evaluate_print(clf_name, y_test, y_test_scores)

average_precision_score(y_test, y_test_scores)

with open("../data/lsa.npy", "wb") as f:
    np.save(f, X_text_train)
    np.save(f, X_text_test)

# save the model
dump(clf_text, '../res/xgbod_text.joblib')

# Combine text and numerical features

X_train_all = np.hstack([X_pca_train, X_train_scaled.to_numpy()])
X_test_all = np.hstack([X_pca_test, X_test_scaled.to_numpy()])

clf_name = 'XGBOD'
clf_all = XGBOD(random_state=42)
clf_all.fit(X_train_all, y_train)

# get the prediction labels and outlier scores of the training data
y_train_pred = clf_all.labels_  # binary labels (0: inliers, 1: outliers)
y_train_scores = clf_all.decision_scores_  # raw outlier scores

# get the prediction on the test data
y_test_pred = clf_all.predict(X_test_all)  # outlier labels (0 or 1)
y_test_scores = clf_all.decision_function(X_test_all)  # outlier scores

# evaluate and print the results
print("\nOn Training Data:")
evaluate_print(clf_name, y_train, y_train_scores)
print("\nOn Test Data:")
evaluate_print(clf_name, y_test, y_test_scores)

average_precision_score(y_test, y_test_scores)

# LDA

lda = LatentDirichletAllocation(
    n_components=16,
    max_iter=5,
    learning_method="online",
    learning_offset=50.0,
    random_state=0,
)

X_lda_train = lda.fit_transform(X_text_train)
X_lda_test = lda.transform(X_text_test)


with open("../res/lda.npy", "wb") as f:
    np.save(f, X_lda_train)
    np.save(f, X_lda_test)


# Model interpretation

# most important words for each topic
vocab = vectorizer.get_feature_names_out()

for i, comp in enumerate(svd.components_):
    vocab_comp = zip(vocab, comp)
    sorted_words = sorted(vocab_comp, key= lambda x:x[1], reverse=True)[:10]
    print("Topic "+str(i)+": ")
    for t in sorted_words:
        print(t[0],end=" ")
    print("\n")

y_test_scores

# SOTA anomally detection models

# ECOD
clf_name = 'ECOD'
clf = ECOD()

# you could try parallel version as well.
# clf = ECOD(n_jobs=2)
clf.fit(X_pca_train)

# get the prediction labels and outlier scores of the training data
y_train_pred = clf.labels_  # binary labels (0: inliers, 1: outliers)
y_train_scores = clf.decision_scores_  # raw outlier scores

# get the prediction on the test data
y_test_pred = clf.predict(X_pca_test)  # outlier labels (0 or 1)
y_test_scores = clf.decision_function(X_pca_test)  # outlier scores

# evaluate and print the results
print("\nOn Training Data:")
evaluate_print(clf_name, y_train, y_train_scores)
print("\nOn Test Data:")
evaluate_print(clf_name, y_test, y_test_scores)

clf_name = 'DIF'
clf = DIF()
clf.fit(X_pca_train)

# get the prediction labels and outlier scores of the training data
y_train_pred = clf.labels_  # binary labels (0: inliers, 1: outliers)
y_train_scores = clf.decision_scores_  # raw outlier scores

# get the prediction on the test data
y_test_pred = clf.predict(X_pca_test)  # outlier labels (0 or 1)
y_test_scores = clf.decision_function(X_pca_test)  # outlier scores

# evaluate and print the results
print("\nOn Training Data:")
evaluate_print(clf_name, y_train, y_train_scores)
print("\nOn Test Data:")
evaluate_print(clf_name, y_test, y_test_scores)

# LightGBM

clf = lgb.LGBMClassifier()
clf.fit(X_pca_train, y_train)
y_pred = clf.predict(X_pca_test)
average_precision_score(y_test, y_pred)
